Market_Mind/src/app.py
import streamlit as st
import yfinance as yf
import pandas as pd
import numpy as np
import plotly.graph_objects as go
from plotly.subplots import make_subplots
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import finnhub
import nltk
from datetime import datetime, timedelta

# ---------------- INITIALISATION ----------------
try:
    nltk.data.find("vader_lexicon")
except LookupError:
    nltk.download("vader_lexicon")

st.set_page_config(page_title="Aurora 2026 | Quant Terminal", layout="wide")

# ---------------- TERMINAL UI CONFIGURATION ----------------
st.markdown("""
<style>
    .main { background-color: #0b0e14; color: #c9d1d9; font-family: 'Courier New', Courier, monospace; }
    [data-testid="stSidebar"] { background-color: #161b22 !important; border-right: 1px solid #30363d; }
    div[data-testid="stMetric"] { background-color: #0d1117; border: 1px solid #30363d; padding: 10px; border-radius: 4px; }
    .stTabs [data-baseweb="tab-list"] { background-color: #0b0e14; border-bottom: 1px solid #30363d; }
    .stTabs [aria-selected="true"] { background-color: #1f6feb !important; color: white !important; }
    .stButton>button { width: 100%; border-radius: 4px; border: 1px solid #30363d; background: #21262d; color: #c9d1d9; }
</style>
""", unsafe_allow_html=True)

# ---------------- SIDEBAR CONTROLS ----------------
with st.sidebar:
    st.title("FINOVA X ISTE")
    ticker = st.text_input("SYMBOL TICKER", value="AAPL").upper()
    time_period = st.selectbox("HISTORY RANGE", ["1y", "2y", "5y"])
    
    st.divider()
    st.write("**DATA CONFIGURATION**")
    finnhub_key = st.text_input("FINNHUB API KEY", type="password")
    
    st.divider()
    st.write("**RISK PARAMETERS**")
    stop_loss = st.slider("STOP LOSS %", 1, 10, 5) / 100
    take_profit = st.slider("TAKE PROFIT %", 5, 20, 10) / 100

# ---------------- DATA PIPELINE ----------------
@st.cache_data
def fetch_market_data(symbol, period):
    # TODO: Implement yfinance download logic
    # TODO: Clean multi-index headers using get_level_values(0)
    data = pd.DataFrame() 
    return data

# ---------------- QUANT ENGINE ----------------
def apply_technical_indicators(data):
    # TODO: Calculate Relative Strength Index (RSI)
    
    # TODO: Construct Bollinger Bands (Upper, Lower, and 20-day SMA)

    # TODO: Implement MACD Line and Signal Line (EMA12 - EMA26)
    
    return data

# ---------------- SENTIMENT ENGINE ----------------
def get_mood_score(symbol, api_key):
    sid = SentimentIntensityAnalyzer()
    # TODO: Fetch 10 headlines from Finnhub within 30-day window
    # TODO: Map VADER compound scores to headlines and calculate average
    return 0, pd.DataFrame(columns=["Headline", "Compound"])

# ---------------- VERDICT ENGINE ----------------
def get_final_verdict(data, sentiment):
    # TODO: Define Decision Logic (e.g., Score +2 for RSI Oversold)
    # TODO: Implement MACD Crossover detection (prev vs last)
    # TODO: Return formatted Verdict String, Terminal Color, and Reason list
    score = 0
    reasons = ["LOGIC PENDING IMPLEMENTATION"]
    return "NEUTRAL", "#ffd700", reasons, score

# ---------------- BACKTEST LAB (STABLE INFRASTRUCTURE) ----------------
def run_backtest(data, sl, tp):
    bt_df = data.copy()
    cash, pos = 10000.0, 0.0
    trades = []
    
    for i in range(1, len(bt_df)):
        curr, prev = bt_df.iloc[i], bt_df.iloc[i-1]
        
        # Internal Logic check
        buy_sig = (prev['MACD'] <= prev['MACD_Signal'] and curr['MACD'] > curr['MACD_Signal']) and curr['RSI'] < 50
        
        if buy_sig and cash > 0:
            pos = cash / curr['Close']
            entry_p = curr['Close']
            cash = 0
            trades.append({'entry_date': bt_df.index[i], 'entry_p': entry_p})
        
        elif pos > 0:
            change = (curr['Close'] - entry_p) / entry_p
            if change <= -sl or change >= tp or curr['RSI'] > 70:
                cash = pos * curr['Close']
                pnl = (curr['Close'] - entry_p) * pos
                pos = 0
                trades[-1].update({'exit_date': bt_df.index[i], 'exit_p': curr['Close'], 'pnl': pnl})

        bt_df.at[bt_df.index[i], 'Total_Value'] = cash + (pos * curr['Close'])
    
    return bt_df, pd.DataFrame([t for t in trades if 'pnl' in t])

# ---------------- APP ORCHESTRATION ----------------
df = fetch_market_data(ticker, time_period)
if not df.empty:
    df = apply_technical_indicators(df)
    avg_s, news = get_mood_score(ticker, finnhub_key)
    v_str, v_col, v_reasons, v_score = get_final_verdict(df, avg_s)

    t1, t2, t3, t4, t5 = st.tabs(["MARKET", "ENGINE", "MOOD", "VERDICT", "BACKTEST"])

    with t1:
        fig = make_subplots(rows=2, cols=1, shared_xaxes=True)
        fig.add_trace(go.Candlestick(x=df.index, open=df['Open'], high=df['High'], low=df['Low'], close=df['Close'], name="Price"), row=1, col=1)
        fig.update_layout(template="plotly_dark", height=600, xaxis_rangeslider_visible=False)
        st.plotly_chart(fig, use_container_width=True)

    with t4:
        st.markdown(f"<h1 style='text-align:center;color:{v_col};'>{v_str}</h1>", unsafe_allow_html=True)
        for r in v_reasons: st.code(r, language="bash")

    with t5:
        if st.button("EXECUTE HISTORICAL SIMULATION"):
            res, history = run_backtest(df, stop_loss, take_profit)
            st.metric("FINAL EQUITY", f"${res['Total_Value'].iloc[-1]:,.2f}")
            st.line_chart(res['Total_Value'])
            st.dataframe(history)