Market_Mind/requirements.txt
yfinance
pandas
numpy
matplotlib
plotly
nltk
finnhub-python